Stop RsJs from intecepting key strokes sent to the notification input.
RxJs previously intercepted both arrow keys and the  spacebar to move
the watch view and handle the face button press respectivley. This causes
issue when the text input of the notifcation creator uses them. This
commit prevents these keys from being silenced when the text input id is in
focus.